profile-account: RadioGroupInput - unnecessary onChange handler removed - unnecessary value check removed - unnecessary RadioInput Component removed - RadioGroup renamed to RadioGroupInput and moved to inputs folder
